requirements.txt
asgiref==3.7.2
Django==5.0
sqlparse==0.4.4

djangorestframework~=3.14.0